csv-kafka-producer/requirements.txt
# Для локального запуска вне Docker: pip install -r requirements.txt
kafka-python>=2.0.2,<3
